Fixed a precision error in vector2d, fixed object rendering cycle

+++ TowerBloonsDefenseKlon/TowerBloonsDefenseKlon/Vector2D.cpp
@@ -19,6 +19,6 @@ void Vector2D::divide(double value) {
 }
 
 void Vector2D::multiply(double value) {
-	this->x = (double)this->x * value;
-	this->y = (double)this->y * value;
+	this->x = this->x * value;
+	this->y = this->y * value;
 }
